Final addition to 0.5.0
We're halfway there! Now seriously, I added a system that will generate special strings used for storing the track's layout and I also added a way to decode this string. It is a relatively small addition but it's pretty complex and overall important to the final steps of this project.