Major structure rehaul
I've redone the structure a bit. The individual classes are finally split to their own files, so the code is more readable. This has been in my mind since I started this project, but some features made it difficult as it refused to cooperate. I will work on splitting even more functions in the future