Added script for Kalman filter
I'm trying out to implement a simple Kalman filter for the GPS so I could get more stable routes out of it, however even the simple algorithm in this case is pretty complex and that's not even half of the problem as I have to implement it into a such large bundle of code. This will take some time to be fully working.